BUGFIX #1 : Player block Collisions
All Direct Collisions with blocks are functional.
2 BUGs left:

1- When the player is on the edge of a block and going towards it; the player's position will change according to the edge position (opposite axes of movement). Making so the player is thrown out of the block and then continues moving according to the first movement clicked by the user.

E.g.: If the player is on top of the left edge of a block and the player is going down, The player will slide to the left until being out of the block, then go down again (while still holding the arrow down).

2- When the player goes in a diagonal, colliding with two blocks in both axes, the player will pass through in a diagonal.